more merging in tests and functionality
