+++ lambdas/dla-migration/package.json
@@ -26,11 +26,9 @@
     "@cumulus/cumulus-message-adapter-js": "2.0.5",
     "@cumulus/message": "18.2.0",
     "@cumulus/types": "18.2.0",
-    "@types/minimist": "^1.2.5",
     "lodash.clone": "^4.5.0",
     "lodash.get": "^4.4.2",
     "lodash.zip": "^4.2.0",
-    "minimist": "^1.2.8",
     "p-retry": "^4.2.0",
     "pmap": "^2.0.0",
     "tsc": "^2.0.4"
